+++ src/common.ts
@@ -1,16 +1,11 @@
 import axios from 'axios';
-import fs from 'fs';
+import { mkdirSync, createWriteStream } from 'fs';
+import path from 'path';
 
-export interface ArchUrl {
-    arm64: string;
-    x64: string;
-}
-
-export interface JLinkVariant {
-    linux: ArchUrl;
-    darwin: ArchUrl;
-    win32: ArchUrl;
-}
+export const platforms = ['darwin', 'linux', 'win32'] as const;
+export const archs = ['arm64', 'x64'] as const;
+export type ArchUrl = Record<(typeof archs)[number], string>;
+export type JLinkVariant = Record<(typeof platforms)[number], ArchUrl>;
 
 export interface JLinkIndex {
     version: string;
@@ -47,9 +42,10 @@ export const fetchIndex = async () => {
 export const saveToFile = (
     stream: NodeJS.ReadableStream,
     destinationFile: string,
-): Promise<string> =>
-    new Promise((resolve, reject) => {
-        const file = fs.createWriteStream(destinationFile);
+): Promise<string> => {
+    mkdirSync(path.dirname(destinationFile), { recursive: true });
+    return new Promise((resolve, reject) => {
+        const file = createWriteStream(destinationFile);
         stream.pipe(file);
         stream.on('error', reject);
         stream.on('end', () => {
@@ -58,3 +54,4 @@ export const saveToFile = (
             });
         });
     });
+};